address review feedback; split up extension controllers in separate packages
