+++ cmd/gardener-extension-shoot-cert-service/app/options.go
@@ -11,7 +11,7 @@ import (
 	heartbeatcmd "github.com/gardener/gardener/extensions/pkg/controller/heartbeat/cmd"
 
 	certificateservicecmd "github.com/gardener/gardener-extension-shoot-cert-service/pkg/cmd"
-	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension"
+	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension/shared"
 )
 
 // ExtensionName is the name of the extension.
@@ -19,16 +19,17 @@ const ExtensionName = "extension-shoot-cert-service"
 
 // Options holds configuration passed to the Certificate Service controller.
 type Options struct {
-	generalOptions     *controllercmd.GeneralOptions
-	certOptions        *certificateservicecmd.CertificateServiceOptions
-	restOptions        *controllercmd.RESTOptions
-	managerOptions     *controllercmd.ManagerOptions
-	controllerOptions  *controllercmd.ControllerOptions
-	healthOptions      *controllercmd.ControllerOptions
-	heartbeatOptions   *heartbeatcmd.Options
-	controllerSwitches *controllercmd.SwitchOptions
-	reconcileOptions   *controllercmd.ReconcilerOptions
-	optionAggregator   controllercmd.OptionAggregator
+	generalOptions                *controllercmd.GeneralOptions
+	certOptions                   *certificateservicecmd.CertificateServiceOptions
+	restOptions                   *controllercmd.RESTOptions
+	managerOptions                *controllercmd.ManagerOptions
+	shootControllerOptions        *controllercmd.ControllerOptions
+	controlPlaneControllerOptions *controllercmd.ControllerOptions
+	healthOptions                 *controllercmd.ControllerOptions
+	heartbeatOptions              *heartbeatcmd.Options
+	controllerSwitches            *controllercmd.SwitchOptions
+	reconcileOptions              *controllercmd.ReconcilerOptions
+	optionAggregator              controllercmd.OptionAggregator
 }
 
 // NewOptions creates a new Options instance.
@@ -41,11 +42,15 @@ func NewOptions() *Options {
 			// These are default values.
 			LeaderElection:          true,
 			LeaderElectionID:        controllercmd.LeaderElectionNameID(ExtensionName),
-			LeaderElectionNamespace: os.Getenv(extension.EnvLeaderElectionNamespace),
+			LeaderElectionNamespace: os.Getenv(shared.EnvLeaderElectionNamespace),
 		},
-		controllerOptions: &controllercmd.ControllerOptions{
+		shootControllerOptions: &controllercmd.ControllerOptions{
 			// This is a default value.
 			MaxConcurrentReconciles: 5,
+		},
+		controlPlaneControllerOptions: &controllercmd.ControllerOptions{
+			// This is a default value.
+			MaxConcurrentReconciles: 2,
 		},
 		healthOptions: &controllercmd.ControllerOptions{
 			// This is a default value.
@@ -55,7 +60,7 @@ func NewOptions() *Options {
 			// This is a default value.
 			ExtensionName:        ExtensionName,
 			RenewIntervalSeconds: 30,
-			Namespace:            os.Getenv(extension.EnvLeaderElectionNamespace),
+			Namespace:            os.Getenv(shared.EnvLeaderElectionNamespace),
 		},
 		controllerSwitches: certificateservicecmd.ControllerSwitches(),
 		reconcileOptions:   &controllercmd.ReconcilerOptions{},
@@ -65,7 +70,8 @@ func NewOptions() *Options {
 		options.generalOptions,
 		options.restOptions,
 		options.managerOptions,
-		options.controllerOptions,
+		options.shootControllerOptions,
+		controllercmd.PrefixOption("controlplane-", options.controlPlaneControllerOptions),
 		options.certOptions,
 		controllercmd.PrefixOption("healthcheck-", options.healthOptions),
 		controllercmd.PrefixOption("heartbeat-", options.heartbeatOptions),
