+++ pkg/cmd/options.go
@@ -20,7 +20,8 @@ import (
 	config "github.com/gardener/gardener-extension-shoot-cert-service/pkg/apis/config"
 	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/apis/config/v1alpha1"
 	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/apis/config/validation"
-	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension"
+	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension/controlplane"
+	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension/shoot"
 	healthcheckcontroller "github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/healthcheck"
 )
 
@@ -93,8 +94,8 @@ func (c *CertificateServiceConfig) Apply(config *config.Configuration) {
 // ControllerSwitches are the cmd.SwitchOptions for the provider controllers.
 func ControllerSwitches() *cmd.SwitchOptions {
 	return cmd.NewSwitchOptions(
-		cmd.Switch(extension.ControllerName, extension.AddToManager),
-		cmd.Switch(extension.ControllerName2, extension.AddToManager2),
+		cmd.Switch(shoot.ControllerName, shoot.AddToManager),
+		cmd.Switch(controlplane.ControllerName, controlplane.AddToManager),
 		cmd.Switch(extensionshealthcheckcontroller.ControllerName, healthcheckcontroller.AddToManager),
 		cmd.Switch(extensionsheartbeatcontroller.ControllerName, extensionsheartbeatcontroller.AddToManager),
 	)
