+++ pkg/controller/extension/controlplane/add_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package extension
+package controlplane
 
 import (
 	"context"
