+++ pkg/controller/extension/shared/assets/cert-dashboard.json
