+++ pkg/controller/extension/shared/assets/crd-cert.gardener.cloud_certificaterevocations.yaml
