+++ pkg/controller/extension/shared/assets/crd-cert.gardener.cloud_certificates.yaml
