+++ pkg/controller/extension/shared/assets/doc.go
