+++ pkg/controller/extension/shared/deployer_suite_test.go
@@ -1,0 +1,13 @@
+package shared
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestDNSManagement(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Extension Shared Deployer Suite")
+}
