--- pkg/controller/extension/testassets/standard_shoot.yaml
@@ -1,207 +1,0 @@
-
----
-# Source: shoot-cert-management-shoot/templates/crds-v1.yaml
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  name: certificates.cert.gardener.cloud
-  labels:
-    shoot.gardener.cloud/no-cleanup: "true"
-    app.kubernetes.io/name: gardener-extension-shoot-cert-service
-    app.kubernetes.io/instance: shoot-cert-management-shoot
-    app.kubernetes.io/managed-by: Helm
-spec:
-  group: cert.gardener.cloud
-  names:
-    kind: Certificate
-    listKind: CertificateList
-    plural: certificates
-    shortNames:
-      - cert
-    singular: certificate
-  scope: Namespaced
-  versions: [] # placeholder for comparison
----
-# Source: shoot-cert-management-shoot/templates/crds-v1.yaml
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  name: certificaterevocations.cert.gardener.cloud
-  labels:
-    shoot.gardener.cloud/no-cleanup: "true"
-    app.kubernetes.io/name: gardener-extension-shoot-cert-service
-    app.kubernetes.io/instance: shoot-cert-management-shoot
-    app.kubernetes.io/managed-by: Helm
-spec:
-  group: cert.gardener.cloud
-  names:
-    kind: CertificateRevocation
-    listKind: CertificateRevocationList
-    plural: certificaterevocations
-    shortNames:
-      - certrevoke
-    singular: certificaterevocation
-  scope: Namespaced
-  versions: [] # placeholder for comparison
----
-# Source: shoot-cert-management-shoot/templates/rbac.yaml
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: extensions.gardener.cloud:extension-shoot-cert-service:shoot
-  labels:
-    app.kubernetes.io/instance: shoot-cert-management-shoot
-    app.kubernetes.io/managed-by: Helm
-rules:
-  - apiGroups:
-      - "networking.k8s.io"
-    resources:
-      - ingresses
-    verbs:
-      - get
-      - list
-      - update
-      - watch
-  - apiGroups:
-      - "gateway.networking.k8s.io"
-    resources:
-      - gateways
-      - httproutes
-    verbs:
-      - get
-      - list
-      - update
-      - watch
-  - apiGroups:
-      - "networking.istio.io"
-    resources:
-      - gateways
-      - virtualservices
-    verbs:
-      - get
-      - list
-      - update
-      - watch
-  - apiGroups:
-      - ""
-    resources:
-      - services
-    verbs:
-      - get
-      - list
-      - update
-      - watch
-  - apiGroups:
-      - ""
-    resources:
-      - secrets
-    verbs:
-      - get
-      - list
-      - update
-      - watch
-      - create
-      - delete
-  - apiGroups:
-      - cert.gardener.cloud
-    resources:
-      - certificates
-      - certificates/status
-      - certificaterevocations
-      - certificaterevocations/status
-    verbs:
-      - get
-      - list
-      - update
-      - watch
-      - create
-      - delete
-  - apiGroups:
-      - ""
-    resources:
-      - events
-    verbs:
-      - create
-      - patch
-  - apiGroups:
-      - apiextensions.k8s.io
-    resources:
-      - customresourcedefinitions
-    verbs:
-      - get
-      - list
-      - update
-      - create
-      - watch
----
-# Source: shoot-cert-management-shoot/templates/rbac.yaml
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: extensions.gardener.cloud:extension-shoot-cert-service:shoot
-  labels:
-    app.kubernetes.io/instance: shoot-cert-management-shoot
-    app.kubernetes.io/managed-by: Helm
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: extensions.gardener.cloud:extension-shoot-cert-service:shoot
-subjects:
-  - kind: ServiceAccount
-    name: extension-shoot-cert-service
-    namespace: kube-system
----
-# Source: shoot-cert-management-shoot/templates/cert-management-role.yaml
-apiVersion: rbac.authorization.k8s.io/v1
-kind: Role
-metadata:
-  name: extensions.gardener.cloud:extension-shoot-cert-service:cert-controller-manager
-  namespace: kube-system
-rules:
-  - apiGroups:
-      - ""
-    resources:
-      - configmaps
-    verbs:
-      - create
-  - apiGroups:
-      - ""
-    resources:
-      - configmaps
-    verbs:
-      - get
-      - watch
-      - update
-    resourceNames:
-      - shoot-cert-service
-  - apiGroups:
-      - coordination.k8s.io
-    resources:
-      - leases
-    verbs:
-      - create
-  - apiGroups:
-      - coordination.k8s.io
-    resources:
-      - leases
-    resourceNames:
-      - shoot-cert-service
-    verbs:
-      - get
-      - watch
-      - update
----
-# Source: shoot-cert-management-shoot/templates/cert-management-rolebinding.yaml
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: extensions.gardener.cloud:extension-shoot-cert-service:cert-controller-manager
-  namespace: kube-system
-subjects:
-  - kind: ServiceAccount
-    name: extension-shoot-cert-service
-    namespace: kube-system
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: Role
-  name: extensions.gardener.cloud:extension-shoot-cert-service:cert-controller-manager
