+++ pkg/controller/healthcheck/add.go
@@ -20,7 +20,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 
 	certv1alpha1 "github.com/gardener/gardener-extension-shoot-cert-service/pkg/apis/service/v1alpha1"
-	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension"
+	"github.com/gardener/gardener-extension-shoot-cert-service/pkg/controller/extension/shoot"
 )
 
 var (
@@ -39,7 +39,7 @@ func RegisterHealthChecks(ctx context.Context, mgr manager.Manager, opts healthc
 	}
 
 	return healthcheck.DefaultRegistration(
-		extension.Type,
+		shoot.Type,
 		extensionsv1alpha1.SchemeGroupVersion.WithKind(extensionsv1alpha1.ExtensionResource),
 		func() client.ObjectList { return &extensionsv1alpha1.ExtensionList{} },
 		func() extensionsv1alpha1.Object { return &extensionsv1alpha1.Extension{} },
